TWITTER/requirements.txt
flask-cors == 3.0.7
pyjwt == 1.7.1
flask-pymongo == 2.2.0
cryptography == 2.4.2
pytest == 4.2.1
requests == 2.21.0
flasgger == 0.9.2
apscheduler == 3.6.0
ipython == 7.2.0
